GPU 增量 PageRank 演示成功
- 创建 examples/gpu_demo.rs：演示 GPU 加速的增量 PageRank
- 修复测试逻辑：正确初始化 new_pr（用旧值），只更新受影响顶点
- GPU 结果正确：与 CPU 计算结果一致

验证结果：
- GPU 计算结果: [0.235, 0.2, 0.2, 0.235, 0.2]
- CPU 计算结果: [0.235, 0.2, 0.2, 0.235, 0.2]
- ✓ GPU 结果正确！

下一步：
- 更新 src/incremental_pagerank.rs 使用正确的 GPU 内核
- 更新 src/gpu/incremental_pagerank.metal 实现完整的增量 PageRank 算法

+++ src/gpu/mod.rs
@@ -82,13 +82,13 @@ impl GPUAccelerator {
         encoder.set_buffer(0, Some(&affected_buffer), 0);
         
         let affected_count_u32 = affected_count as u32;
-        encoder.set_bytes(1, std::mem::size_of::<u32>(), &affected_count_u32 as *const u32 as *const c_void);
+        encoder.set_bytes(1, std::mem::size_of::<u32>() as u64, &affected_count_u32 as *const u32 as *const c_void);
         
         encoder.set_buffer(2, Some(&offsets_buffer), 0);
         encoder.set_buffer(3, Some(&targets_buffer), 0);
         encoder.set_buffer(4, Some(&pr_buffer), 0);
         encoder.set_buffer(5, Some(&new_pr_buffer), 0);
-        encoder.set_bytes(6, std::mem::size_of::<u32>(), &vertex_count as *const u32 as *const c_void);
+        encoder.set_bytes(6, std::mem::size_of::<u32>() as u64, &vertex_count as *const u32 as *const c_void);
         
         // Swift 第 334-336 行：调度线程
         let grid_size = MTLSize::new(affected_count, 1, 1);
